% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gallery.R
\name{colorCellsByMeta}
\alias{colorCellsByMeta}
\title{colorCellsByMeta}
\usage{
colorCellsByMeta(
  labelsImgs,
  fullCellMeta,
  cellID_coln = "cell_ID",
  cellLabel_coln = "CellId",
  UID_coln = "UID",
  color_coln = "celltype",
  isDiscrete = TRUE,
  colors_to_use = NULL,
  value_limits = NULL,
  plotresultdir = getwd()
)
}
\arguments{
\item{labelsImgs}{named list of cell label image arrays with one query cell per element, each label image array is a 2D matrix with value in integer indicating cell id}

\item{fullCellMeta}{data.frame for cell metadata of all cells in study}

\item{cellID_coln}{column name of cell_ID in \code{fullCellMeta}}

\item{cellLabel_coln}{column name of cell label values in \code{fullCellMeta}}

\item{UID_coln}{column name of unique identifier for each fov in \code{fullCellMeta}}

\item{color_coln}{column name of metadata used to color cell labels in \code{fullCellMeta}}

\item{isDiscrete}{flag discrete metadata (default = TRUE)}

\item{colors_to_use}{a vector of colors to use; when isDiscrete = TRUE, must have length and names matched to unique values in color_coln}

\item{value_limits}{a vector of two elements for the value range to display, used when isDiscrete = FALSE; default = NULL to use the 1\% and 99\% quantile of values in color_coln across entire study.}

\item{plotresultdir}{output folder for color legend used}
}
\value{
a list of single-frame rgb image array with one element per query cell
}
\description{
color cell labels with cell metadata
}
