% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gallery.R
\name{combineImgByGroup}
\alias{combineImgByGroup}
\title{combineImgByGroup}
\usage{
combineImgByGroup(imgList, groupVector, imgType = c("labels", "morph"))
}
\arguments{
\item{imgList}{named list of image arrays with one query cell per element, names for cell_ID}

\item{groupVector}{named vector of group ID with names to be cell_ID, values to be group ID.}

\item{imgType}{type of input image in \code{imgList}, labels for 2D matrix, morph for multi-channel 3D array}
}
\value{
a nested list of 2 elements, \code{cell_IDs} and \code{combImgObj}, whose elements are organized by group
}
\description{
convert list of image arrays into multi-frame image object, one object per group
}
