% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gallery.R
\name{combineMetaToMorph_imgList}
\alias{combineMetaToMorph_imgList}
\title{combineMetaToMorph_imgList}
\usage{
combineMetaToMorph_imgList(rgbMorphImgs, rgbMetaImgs, morphToShow = c(1, 2, 3))
}
\arguments{
\item{rgbMorphImgs}{list of rgb image arrays with one query cell per element, each rgb image array has at least 3 channels with value in range of 0-1 for RGB color}

\item{rgbMetaImgs}{list of rgb image array with one query cell per element, each rgb image array has exactly 3 channels with zero value for extracellular space}

\item{morphToShow}{a vector of 3 elements for the channel index of each RGB channel in \code{rgbMorphImgs}.}
}
\value{
a multi-frame rgb image array with intracellular region colored by \code{rgbMetaImgs} but extracellular region colored by \code{rgbMorphImgs}
}
\description{
combine 2 set of images into one set
}
