% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{combineMorphList}
\alias{combineMorphList}
\title{combineMorphList}
\usage{
combineMorphList(imgList1, imgList2)
}
\arguments{
\item{imgList1}{named list of image arrays with one query cell per element, names for cell_ID}

\item{imgList2}{named list of image arrays with one query cell per element, names for cell_ID}
}
\value{
a named list of multi-channel image arrays with one query cell per element, names for cell_ID
}
\description{
combine 2 sets of multi-channel image array together and
}
