% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gallery.R
\name{createComposite_multiFrame}
\alias{createComposite_multiFrame}
\title{createComposite_multiFrame}
\usage{
createComposite_multiFrame(
  imgObj,
  vizChannels = NULL,
  channelColors = NULL,
  equalizeFlags = NULL
)
}
\arguments{
\item{imgObj}{array or EBImage class for multi-frame image}

\item{vizChannels}{vector of index of channels to show}

\item{channelColors}{vector of monochrome to use for each channel in vizChannels}

\item{equalizeFlags}{vector of logic values to equalize each channel in vizChannels}
}
\value{
a multi-frame rgb image array of 3 channels in EBImage class
}
\description{
create composite RGB images from the input single-frame multi-channel image array
}
