% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{decodeStitchedLabelVal}
\alias{decodeStitchedLabelVal}
\title{decodeStitchedLabelVal}
\usage{
decodeStitchedLabelVal(labVals)
}
\arguments{
\item{labVals}{vector of encoded label value used in napari-CosMx stitched data}
}
\value{
data.frame with fov and CellId for each encoded label value
}
\description{
decode the stitched label value in napari-CosMx stitched data to fov and CellId
}
