% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gallery.R
\name{displayCellId_singleImg}
\alias{displayCellId_singleImg}
\title{displayCellId_singleImg}
\usage{
displayCellId_singleImg(
  imgLabel,
  query_cellID = NULL,
  cellID_converter = NULL,
  seed = 123
)
}
\arguments{
\item{imgLabel}{single-frame label image for one query cell, in format of a 2D matrix with value in integer indicating cell id}

\item{query_cellID}{cell_ID of input \code{labelsImg}}

\item{cellID_converter}{a named vector of how values in imgLabel is linked to cell_ID, useful when input is based on stitched image; default = NULL to show the label value as cell_ID}

\item{seed}{seed for randomized colors of cell_ID}
}
\value{
a named vector of colors used for plotting cell labels with names for cell id in integer value
}
\description{
display cell_id of one single-frame label image, the identity of id for each label is shown as color in legend
}
