% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{encodeCellId}
\alias{encodeCellId}
\title{encodeCellId}
\usage{
encodeCellId(cellMeta, cellLabel_coln = "CellId", fov_coln = "fov")
}
\arguments{
\item{cellMeta}{data.frame of cell metadata}

\item{cellLabel_coln}{column name of cell label values at FOV level in \code{cellMeta} (default = 'CellId')}

\item{fov_coln}{column name of fov in \code{cellMeta}}
}
\value{
a vector of encoded label value used in napari-CosMx stitched data
}
\description{
encode fov and CellId to the stitched label value used in napari-CosMx stitched data
}
