% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extractNeighborhoodImgs}
\alias{extractNeighborhoodImgs}
\title{extractNeighborhoodImgs}
\usage{
extractNeighborhoodImgs(
  pixel_size_um = 0.120280945,
  cropSize_um = 50,
  cells_to_use,
  morphFileDF,
  spatColns = c("CenterX", "CenterY"),
  equalizeFOV_flag = FALSE,
  coreNum = 5,
  label_bit = c(16, 8, 32)
)
}
\arguments{
\item{pixel_size_um}{pixel size in um unit of raw images}

\item{cropSize_um}{dimension of extracted neighborhood for query cells in um unit}

\item{cells_to_use}{vector of cell_ID for query cells in \verb{c_[slide]_[fov]_[CellId]} format}

\item{morphFileDF}{data.frame of file path for the paired cell segmentation outputs of each fov}

\item{spatColns}{a vector of column names for xy spatial coordinates of each cell in Cell_Stats files, in pixel unit}

\item{equalizeFOV_flag}{flag to equalize intensity of morph images for each channel at fov level (default = FALSE)}

\item{coreNum}{number of cores for mclapply parallelization}

\item{label_bit}{encoded range for label image (default = 16)}
}
\value{
a nested list of 2 elements, \code{labels} and \code{morph}, which contains list of cell-wise image array
}
\description{
extract the neighborhood images around query cells
}
