% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extractNeighborhoodImgs_fromProteinDir}
\alias{extractNeighborhoodImgs_fromProteinDir}
\title{extractNeighborhoodImgs_fromProteinDir}
\usage{
extractNeighborhoodImgs_fromProteinDir(
  pixel_size_um = 0.120280945,
  cropSize_um = 50,
  cells_to_use,
  targets_to_load,
  proteinFileDF,
  spatColns = c("CenterX", "CenterY"),
  equalizeFOV_flag = FALSE,
  coreNum = 5
)
}
\arguments{
\item{pixel_size_um}{pixel size in um unit of raw images}

\item{cropSize_um}{dimension of extracted neighborhood for query cells in um unit}

\item{cells_to_use}{vector of cell_ID for query cells in \verb{c_[slide]_[fov]_[CellId]} format}

\item{targets_to_load}{a vector of protein targets to load}

\item{proteinFileDF}{data.frame of file path for the protein images of each fov}

\item{spatColns}{a vector of column names for xy spatial coordinates of each cell in PerCellStats files, in pixel unit}

\item{equalizeFOV_flag}{flag to equalize intensity of morph images for each channel at fov level (default = FALSE)}

\item{coreNum}{number of cores for mclapply parallelization}
}
\value{
a nested list of 2 elements, \code{morph}and \code{morphChannelIDs}, the former of which contains cell-wise image array with multi-channel for different protein targets, could be appended to the \code{morph} outputs of \code{extractNeighborhoodImgs} function.
}
\description{
extract the neighborhood images around query cells for ProteinImages
}
