% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extractNeighborhoodImgs_fromStitched}
\alias{extractNeighborhoodImgs_fromStitched}
\title{extractNeighborhoodImgs_fromStitched}
\usage{
extractNeighborhoodImgs_fromStitched(
  pixel_size_um = 0.120280945,
  cropSize_um = 50,
  cells_to_use,
  zarrDF,
  global_size,
  cellMeta,
  cellID_coln = "cell_ID",
  cellLabel_coln = "labVal",
  spatColns = c("global_x_px", "global_y_px"),
  flip_axes = TRUE,
  equalize_flag = FALSE,
  coreNum = 5
)
}
\arguments{
\item{pixel_size_um}{pixel size in um unit of raw images}

\item{cropSize_um}{dimension of extracted neighborhood for query cells in um unit}

\item{cells_to_use}{vector of cell_ID for query cells in \verb{c_[slide]_[fov]_[CellId]} format}

\item{zarrDF}{data.frame of file path for the paired napari-CosMx stitched data}

\item{global_size}{a vector for the dimension of stitched image in pixel unit and c(column/width, rows/height) order, output of \code{prepStitchedData} or \code{getGlobalCoords_forStitched} function}

\item{cellMeta}{data.frame for cell metadata of cells in napari-CosMx stitched data, output of \code{prepStitchedData} or \code{getGlobalCoords_forStitched} function}

\item{cellID_coln}{column name of cell_ID in \code{cellMeta}}

\item{cellLabel_coln}{column name of cell label values in \code{cellMeta} (default = 'labVal')}

\item{spatColns}{a vector of column names for global xy spatial coordinates of each cell in \code{cellMeta}, in pixel unit and c(column index, row index) order}

\item{flip_axes}{flag to flip between row and column axes of extracted image upon return (default = TRUE)}

\item{equalize_flag}{flag to equalize intensity of morph images for each channel at each query cell level (default = FALSE)}

\item{coreNum}{number of cores for mclapply parallelization}
}
\value{
a nested list of 3 elements, \code{labels}, \code{morph}, \code{morphChannelIDs}, the former of which contain list of cell-wise image array, the 3rd element is a vector of channelID in \code{morph}
}
\description{
extract the neighborhood images around query cells from napari-CosMx stitched data
}
