% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{getCellSegFiles}
\alias{getCellSegFiles}
\title{getCellSegFiles}
\usage{
getCellSegFiles(CellStatsDir, slideID = 1)
}
\arguments{
\item{CellStatsDir}{file path to CellStatsDir folder}

\item{slideID}{slide ID that would be assigned to cells in the query flow cell}
}
\value{
data.frame of file path for each fov's CellLabel, morphology 2D and cell stats files
}
\description{
collect cell segmentation output file paths for CosMx assay
}
