% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{getProteinDirFiles}
\alias{getProteinDirFiles}
\title{getProteinDirFiles}
\usage{
getProteinDirFiles(ProteinDir, slideID = 1)
}
\arguments{
\item{ProteinDir}{file path to ProteinDir folder}

\item{slideID}{slide ID that would be assigned to cells in the query flow cell}
}
\value{
data.frame of file path for each fov's decoded protein images, and per cell stats files
}
\description{
collect protein image output file paths for CosMx assay
}
