% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gallery.R
\name{overlayCellBorder_imgObj}
\alias{overlayCellBorder_imgObj}
\title{overlayCellBorder_imgObj}
\usage{
overlayCellBorder_imgObj(
  rgbMorphObj,
  labelsObj,
  cell_ids = NULL,
  morphToShow = c(1, 2, 3),
  border_color = "#FFFFFF",
  export_plots = c("none", "morph", "labels", "overlay"),
  plotresultdir = getwd(),
  fileprefix = "group1"
)
}
\arguments{
\item{rgbMorphObj}{multi-frame rgb image array with one query cell per frame, each frame has at least 3 channels with values in range of 0-1 for RGB color}

\item{labelsObj}{multi-frame cell label image array with one query cell per frame, each frame is a 2D matrix with value in integer indicating cell id}

\item{cell_ids}{a vector of cell_id for each frame in input image obj (default = NULL)}

\item{morphToShow}{a vector of 3 elements for the channel index of each RGB channel in \code{rgbMorphImgs}.}

\item{border_color}{hex code for color of cell borders on overlay (default = '#FFFFFF')}

\item{export_plots}{a vector of different output images that would be exported to png file on disk}

\item{plotresultdir}{output folder for plot results}

\item{fileprefix}{file name prefix for output pngs}
}
\value{
a multi-frame rgb image array with cell border overlay on top of morph
}
\description{
overlay cell borders on top of RGB images which could be outputs of \code{\link{createComposite_multiFrame}}, each input imageObj has multiple frames
}
