% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{prepStitchedData}
\alias{prepStitchedData}
\title{prepStitchedData}
\usage{
prepStitchedData(napari_imgDir, slideID = 1)
}
\arguments{
\item{napari_imgDir}{file path to stitched images generated by napari}

\item{slideID}{slide ID that would be assigned to cells in the query flow cell}
}
\value{
a list of 3 data.frame and 1 list
\itemize{
\item{zarrDF}{data.frame for file path of different image channels under \code{napari_imgDir}}
\item{fovOffsets}{data.frame for fov offset used to arrange and stitch FOVs in \code{napari_imgDir}}
\item{cellMeta}{data.frame for cell metadata, linking original fov and CellId to the values in stitched labels image of \code{napari_imgDir}}
\item{dimSetup}{list of 5 elements, \code{fov_size},  \code{pixel_size_mm}, \code{isDASH}, \code{global_size}, \code{topLeft_mm}}
}
}
\description{
collect information for the stitched zarr images generated by napari-CosMx plugin, link the values in labels image to the input FOV and cell_label; slower than \code{getGlobalCoords_forStitched}
}
